Resolving double-close error caught by fdsan.
Stopping the VPN on Android 11+ devices would cause a crash due to fdsan error.  The issue was that we were closing the vpn file descriptor twice.  It was being closed once by "Processor" and again by the vpn.  To fix this issue, we no longer close the file descriptor in the vpn.